Remove support for the old dispatching
