+++ Cargo.toml
@@ -13,16 +13,13 @@ exclude = ["media"]
 
 [features]
 # FIXME: remove "cache-me" that was added by mistake here
-default = ["native-tls", "ctrlc_handler", "teloxide-core/default",  "auto-send", "dispatching2", "cache-me"]
-
-dispatching2 = ["dptree", "cache-me"]
+default = ["native-tls", "ctrlc_handler", "teloxide-core/default",  "auto-send", "cache-me"]
 
 sqlite-storage = ["sqlx"]
 redis-storage = ["redis"]
 cbor-serializer = ["serde_cbor"]
 bincode-serializer = ["bincode"]
 
-frunk- = ["frunk"]
 macros = ["teloxide-macros"]
 
 ctrlc_handler = ["tokio/signal"]
@@ -44,7 +41,6 @@ full = [
     "redis-storage",
     "cbor-serializer",
     "bincode-serializer",
-    "frunk",
     "macros",
     "ctrlc_handler",
     "teloxide-core/full",
@@ -59,28 +55,26 @@ full = [
 
 [dependencies]
 teloxide-core = { version = "0.4", default-features = false }
-teloxide-macros = { version = "0.5.1", optional = true }
+teloxide-macros = { git = "https://github.com/teloxide/teloxide-macros.git", rev = "144eb73aaf39145bf8f6b57eec5c76730961c2f1", optional = true }
 
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
 
-dptree = { version = "0.1.0", optional = true }
+dptree = { version = "0.1.0" }
 
 tokio = { version = "1.8", features = ["fs"] }
 tokio-util = "0.6"
-tokio-stream = "0.1"
 
-flurry = "0.3"
 log = "0.4"
 bytes = "1.0"
 mime = "0.3"
 
 derive_more = "0.99"
 thiserror = "1.0"
-async-trait = "0.1"
 futures = "0.3.15"
 pin-project = "1.0"
 serde_with_macros = "1.4"
+aquamarine = "0.1.11"
 
 sqlx = { version = "0.5", optional = true, default-features = false, features = [
         "runtime-tokio-native-tls",
@@ -90,15 +84,11 @@ sqlx = { version = "0.5", optional = true, default-features = false, features = 
 redis = { version = "0.20", features = ["tokio-comp"], optional = true }
 serde_cbor = { version = "0.11", optional = true }
 bincode = { version = "1.3", optional = true }
-frunk = { version = "0.4", optional = true }
-aquamarine = "0.1.11"
 
 [dev-dependencies]
-smart-default = "0.6.0"
 rand = "0.8.3"
 pretty_env_logger = "0.4.0"
 once_cell = "1.9.0"
-lazy_static = "1.4.0"
 anyhow = "1.0.52"
 serde = "1"
 serde_json = "1"
@@ -106,6 +96,7 @@ tokio = { version = "1.8", features = ["fs", "rt-multi-thread", "macros"] }
 warp = "0.3.0"
 reqwest = "0.10.4"
 chrono = "0.4"
+tokio-stream = "0.1"
 
 [package.metadata.docs.rs]
 all-features = true
@@ -149,5 +140,5 @@ name = "admin"
 required-features = ["macros"]
 
 [[example]]
-name = "dispatching2_features"
+name = "dispatching_features"
 required-features = ["macros"]
