+++ examples/dispatching_features.rs
@@ -1,12 +1,10 @@
 // This example provide a quick overview of the new features in the
-// `dispatching2` module.
+// `dispatching` module.
 
 use rand::Rng;
 
-// You need to import `prelude2` because `prelude` contains items from the old
-// dispatching system, which will be deprecated in the future.
 use teloxide::{
-    prelude2::*,
+    prelude::*,
     types::{Dice, Update},
     utils::command::BotCommand,
 };
@@ -14,7 +12,7 @@ use teloxide::{
 #[tokio::main]
 async fn main() {
     pretty_env_logger::init();
-    log::info!("Starting dispatching2_features_bot...");
+    log::info!("Starting dispatching_features_bot...");
 
     let bot = Bot::from_env().auto_send();
