+++ src/dispatching/repls/repl.rs
@@ -1,14 +1,11 @@
 use crate::{
-    dispatching::{
-        update_listeners, update_listeners::UpdateListener, Dispatcher, DispatcherHandlerRx,
-        UpdateWithCx,
-    },
+    dispatching::{update_listeners, update_listeners::UpdateListener, UpdateFilterExt},
     error_handlers::{LoggingErrorHandler, OnError},
+    types::Update,
 };
-use futures::StreamExt;
-use std::{fmt::Debug, future::Future, sync::Arc};
-use teloxide_core::{requests::Requester, types::Message};
-use tokio_stream::wrappers::UnboundedReceiverStream;
+use dptree::di::{DependencyMap, Injectable};
+use std::fmt::Debug;
+use teloxide_core::requests::Requester;
 
 /// A [REPL] for messages.
 ///
@@ -22,22 +19,16 @@ use tokio_stream::wrappers::UnboundedReceiverStream;
 /// [REPL]: https://en.wikipedia.org/wiki/Read-eval-print_loop
 /// [`Dispatcher`]: crate::dispatching::Dispatcher
 #[cfg(feature = "ctrlc_handler")]
-pub async fn repl<R, H, Fut, E>(requester: R, handler: H)
+pub async fn repl<R, H, E, Args>(bot: R, handler: H)
 where
-    H: Fn(UpdateWithCx<R, Message>) -> Fut + Send + Sync + 'static,
-    Fut: Future<Output = Result<(), E>> + Send + 'static,
+    H: Injectable<DependencyMap, Result<(), E>, Args> + Send + Sync + 'static,
     Result<(), E>: OnError<E>,
-    E: Debug + Send,
-    R: Requester + Send + Clone + 'static,
+    E: Debug + Send + Sync + 'static,
+    R: Requester + Send + Sync + Clone + 'static,
     <R as Requester>::GetUpdates: Send,
 {
-    let cloned_requester = requester.clone();
-    repl_with_listener(
-        requester,
-        handler,
-        update_listeners::polling_default(cloned_requester).await,
-    )
-    .await;
+    let cloned_bot = bot.clone();
+    repl_with_listener(bot, handler, update_listeners::polling_default(cloned_bot).await).await;
 }
 
 /// Like [`repl`], but with a custom [`UpdateListener`].
@@ -53,44 +44,34 @@ where
 /// [`repl`]: crate::dispatching::repls::repl()
 /// [`UpdateListener`]: crate::dispatching::update_listeners::UpdateListener
 #[cfg(feature = "ctrlc_handler")]
-pub async fn repl_with_listener<'a, R, H, Fut, E, L, ListenerE>(
-    requester: R,
-    handler: H,
-    listener: L,
-) where
-    H: Fn(UpdateWithCx<R, Message>) -> Fut + Send + Sync + 'static,
-    Fut: Future<Output = Result<(), E>> + Send + 'static,
+pub async fn repl_with_listener<'a, R, H, E, L, ListenerE, Args>(bot: R, handler: H, listener: L)
+where
+    H: Injectable<DependencyMap, Result<(), E>, Args> + Send + Sync + 'static,
     L: UpdateListener<ListenerE> + Send + 'a,
     ListenerE: Debug,
     Result<(), E>: OnError<E>,
-    E: Debug + Send,
-    R: Requester + Clone + Send + 'static,
+    E: Debug + Send + Sync + 'static,
+    R: Requester + Clone + Send + Sync + 'static,
 {
-    let handler = Arc::new(handler);
+    use crate::dispatching::Dispatcher;
 
-    Dispatcher::new(requester)
-        .messages_handler(|rx: DispatcherHandlerRx<R, Message>| {
-            UnboundedReceiverStream::new(rx).for_each_concurrent(None, move |message| {
-                let handler = Arc::clone(&handler);
+    // Other update types are of no interest to use since this REPL is only for
+    // messages. See <https://github.com/teloxide/teloxide/issues/557>.
+    let ignore_update = |_upd| Box::pin(async {});
 
-                async move {
-                    handler(message).await.log_on_error().await;
-                }
-            })
-        })
-        .setup_ctrlc_handler()
+    #[allow(unused_mut)]
+    let mut dispatcher =
+        Dispatcher::builder(bot, Update::filter_message().branch(dptree::endpoint(handler)))
+            .default_handler(ignore_update)
+            .build();
+
+    #[cfg(feature = "ctrlc_handler")]
+    dispatcher.setup_ctrlc_handler();
+
+    dispatcher
         .dispatch_with_listener(
             listener,
             LoggingErrorHandler::with_custom_text("An error from the update listener"),
         )
         .await;
-}
-
-#[test]
-fn repl_is_send() {
-    let bot = crate::Bot::new("");
-    let repl = crate::repl(bot, |_| async { crate::respond(()) });
-    assert_send(&repl);
-
-    fn assert_send(_: &impl Send) {}
 }
