+++ src/prelude.rs
@@ -1,27 +1,17 @@
 //! Commonly used items.
 
-#![deprecated(note = "Use dispatching2 instead")]
-#![allow(deprecated)]
-
 pub use crate::{
     error_handlers::{LoggingErrorHandler, OnError},
     respond,
 };
 
 pub use crate::dispatching::{
-    dialogue::{
-        exit, next, DialogueDispatcher, DialogueStage, DialogueWithCx, GetChatId, Transition,
-        TransitionIn, TransitionOut,
-    },
-    Dispatcher, DispatcherHandlerRx, DispatcherHandlerRxExt, UpdateWithCx,
+    dialogue::Dialogue, Dispatcher, HandlerExt as _, MessageFilterExt as _, UpdateFilterExt as _,
 };
-
-#[cfg(feature = "macros")]
-pub use crate::teloxide;
 
 pub use teloxide_core::types::{
     CallbackQuery, ChatMemberUpdated, ChosenInlineResult, InlineQuery, Message, Poll, PollAnswer,
-    PreCheckoutQuery, ShippingQuery,
+    PreCheckoutQuery, ShippingQuery, Update,
 };
 
 #[cfg(feature = "auto-send")]
@@ -30,9 +20,4 @@ pub use crate::adaptors::AutoSend;
 #[doc(no_inline)]
 pub use teloxide_core::prelude::*;
 
-#[cfg(feature = "frunk")]
-pub use crate::utils::UpState;
-
-pub use tokio::sync::mpsc::UnboundedReceiver;
-
-pub use futures::StreamExt;
+pub use dptree::{self, prelude::*};
