--- src/prelude2.rs
@@ -1,26 +1,0 @@
-//! Commonly used items (`dispatching2`).
-
-pub use crate::{
-    error_handlers::{LoggingErrorHandler, OnError},
-    respond,
-};
-
-pub use crate::dispatching2::{
-    dialogue::Dialogue, Dispatcher, HandlerExt as _, MessageFilterExt as _, UpdateFilterExt as _,
-};
-
-#[cfg(feature = "macros")]
-pub use crate::teloxide;
-
-pub use teloxide_core::types::{
-    CallbackQuery, ChatMemberUpdated, ChosenInlineResult, InlineQuery, Message, Poll, PollAnswer,
-    PreCheckoutQuery, ShippingQuery, Update,
-};
-
-#[cfg(feature = "auto-send")]
-pub use crate::adaptors::AutoSend;
-
-#[doc(no_inline)]
-pub use teloxide_core::prelude::*;
-
-pub use dptree::{self, prelude::*};
