+++ src/utils/mod.rs
@@ -4,7 +4,6 @@ pub mod command;
 pub mod html;
 pub mod markdown;
 pub(crate) mod shutdown_token;
-mod up_state;
 
 pub use teloxide_core::net::client_from_env;
