+++ src/utils/shutdown_token.rs
@@ -12,7 +12,7 @@ use tokio::sync::Notify;
 
 use crate::dispatching::update_listeners::UpdateListener;
 
-/// A token which used to shutdown [`Dispatcher`].
+/// A token which used to shutdown [`crate::dispatching::Dispatcher`].
 #[derive(Clone)]
 pub struct ShutdownToken {
     dispatcher_state: Arc<DispatcherState>,
@@ -20,7 +20,7 @@ pub struct ShutdownToken {
 }
 
 /// This error is returned from [`ShutdownToken::shutdown`] when trying to
-/// shutdown an idle [`Dispatcher`].
+/// shutdown an idle [`crate::dispatching::Dispatcher`].
 #[derive(Debug)]
 pub struct IdleShutdownError;
